Change benchmarks script

+++ benchmarks/submit_benchmarks.sh
@@ -2,6 +2,11 @@
 
 PROJECT_ROOT="$PWD"/..
 
+PARTITION="$1"
+if [ -z "$PARTITION" ]; then
+    echo "Usage: $0 <partition>. Available partitions: gpu, ampere."
+    exit 1
+fi
+
 rm *.out
-sbatch --partition=gpu run_benchmarks.sh "$PROJECT_ROOT"
-sbatch --partition=ampere run_benchmarks.sh "$PROJECT_ROOT"
+sbatch --partition="$PARTITION" run_benchmarks.sh "$PROJECT_ROOT"
